requirements.txt
pywifi
matplotlib
opencv-python
Pillow
numpy
tkinter
comtypes